Issue #15: Added support for flask-restplus
Because flask-restplus does some different and strange things with its
error handler, I decided to add some sugar to make things easier. In
particular, the new register_error_handler_with_flask_restplus method
automatically registers error handlers for FlaskBuzz derived errors

Additionally, I deprecated the old error_handler method in favor of a
new build_error_handler method that can take a list of additional
functions that should be called with the error as the argument when the
error is handled. This is convenient if, say, your application needs to
log the error or other sorts of bookkeeping